Separated show.c
